feat: 给系统后台添加用户必须绑定权限角色

+++ src/main/java/cn/xu/infrastructure/persistent/dao/IUserDao.java
@@ -22,7 +22,7 @@ public interface IUserDao {
 
     List<User> selectAdminByPage(@Param("page") int page, @Param("size") int size);
 
-    int insertUser(@Param("user") User user);
+    Long insertUser(@Param("user") User user);
 
     int updateUser(@Param("user") User user);
